requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
yfinance==0.2.54
alpha-vantage==2.3.1
pandas==2.0.3
pyspark==3.4.0
pyro-ppl==1.8.6
mlflow==2.14.1
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.36.0
scikit-learn==1.5.0
prophet==1.1.5
matplotlib==3.9.0
requests==2.32.3
setuptools==74.1.2
plotly==5.22.0